escalate the responsility of loading articles to result view
