+++ src/js/components/result-view.js
@@ -8,9 +8,26 @@ var ArticleList  = require('./article-list');
 var Messages     = require('./search-message');
 var Pagination   = require('./pagination');
 var ArticleActor = require('../actors/article-actor');
+var ArticleStore = require('../stores/article-store');
 
 module.exports = React.createClass({
   mixins: [ History ],
+  _onChange: function() {
+    this.setState({ articles  : ArticleStore.getArticles() });
+    this.setState({ isLoading : false });
+  },
+  getInitialState: function() {
+    return {
+      articles: ArticleStore.getArticles(),
+      isLoading: true
+    };
+  },
+  componentDidMount: function() {
+    ArticleStore.addListener(this._onChange);
+  },
+  componentWillUnmount: function() {
+    ArticleStore.removeListener(this._onChange);
+  },
   componentWillMount: function() {
     if (_.isEmpty(this.props.location.query)) {
       this.history.pushState(null, '/');
@@ -32,7 +49,7 @@ module.exports = React.createClass({
         </div>
         <div className="col-md-9">
           <Messages />
-          <ArticleList />
+          <ArticleList articles={ this.state.articles }/>
           <Pagination history={ this.history } />
          </div>
        </div>
